Split v1 and v2 APIs. Add feature flags for each version and make more dependencies optional.
