+++ src/api/v1/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod extractors;
+mod responses;
+mod routes;
+
+pub use routes::routes;
